# Fuentes et al. algorithm for computing hard part of final exponentiation